Tor Browser: run as a pseudo flatpak
With "pseudo" Flatpak we mean that it's not installed from an actual
Flatpak, but generated at build time, and it doesn't contain the
application files or runtime as usual, which would waste disk
space. Since the application files are installed in the usual Tails
filesystem we create an overlay filesystem with the required bits for
the Flatpak runtime.

The motivation for doing this isn't the Flatpak sandbox (we are happy
with AppArmor and Tor Browser's own sandbox) but proper XDG Desktop
Portal support. The AppArmor confinement we've had for years causes
bad UX for in file dialogs when saving/opening files, but the
FileChooser portal improves this. The FileChooser portal can be used
without Flatpak (by setting GTK_USE_PORTAL=1) but xdg-desktop-portal
doesn't detect that AppArmor is used, so it will access files directly
which then brings us back to AppArmor blocking access. But when the it
detects that the application runs as a Flatpak the "doc portal" is
used to give (indirect) access to the files in a way which is allowed
by the AppArmor profile.

Incidentally, using the Settings Portal solves a few other issues:

- We now have minimize/maximize buttons in the title bar
- The "Large Text" accessibility feature works
- Changed cursor size is applied inside the Tor Browser window

Refs: #10422, #15678, #19266, #19328, #19408, #19572