Tor Browser flatpak: allow Tor Browser access to XDG user dirs
We can consider reverting this once
https://bugzilla.mozilla.org/show_bug.cgi?id=1814851 is fixed.

However, it is only in unrestriced directories like these that
flatpak's --file-forwarding works with HTML documents that reference
other local files (e.g. in a _files directory after using Save Page
As); when opening a HTML document in a restricted directory it will
lack images, CSS etc.

refs: #19255