tpsd: don't mount the filesystem when testing the LUKS passphrase
During UpgradeLUKS we test the passphrase to verify that the system
isn't corrupted. We used to check that by mounting the cleartext
device's filesystem after unlocking, but this causes an error if it is
unmountable due to "normal" filesystem corruption (i.e. not because of
LUKS corruption). It is enough to just verify that the filesystem
exists and is of the expected type, so let's do that instead of having
to deal with all potential problems that mounting can cause.

FWIW, after tails#20634 was fixed the automated test suite's "Manual
filesystem repair is successful" scenario started to fail because the
above issue, and that's how this problem was found.