20-time.sh: don't block execution (refs: tails#20668)
Due to #10589 it is not good if the NM dispatcher hooks can take a
long time to execute, since they will block the queue for new
connection events. Here's an example where that causes problems:

1. Connect to Tor in TCA
2. Immediately when TCA reports success, disconnect the network
3. Observe how Tails still thinks we're connected to Tor,
   i.e. tails-tor-has-bootstrapped-flag-file.service is still active
   so the flag file exists, and tails-tor-has-bootstrapped.target is
   still reached.

The problem is that in 20-time.sh we call `systemctl restart
tails-htpdate-notify-user.service` without `--no-block`; if the
network was up it would exit fairly quickly, but if it is down (as in
the example above) we have to wait for it to time out.