Electrum wrapper: also show donation call to "first time users"
This was requested on tails#17850.

The donation prefilling works even if there is no wallet, it will be
shown once the wallet has been created.

But does it makes sense? Let's first observe that "no wallet" doesn't
necessarily mean that the user will create a brand new (empty) wallet,
since the wizard has options to restore wallets from
seed/hardware/master key, in which case it makes sense to treat it as
if a wallet already existed. Supporting these cases seems like a good
idea to me.

However, since our wrapper cannot foretell the future it cannot
distinguish these cases from the "brand new wallet" case, so we have
to live with it too. But it isn't so bad: once the brand new wallet is
crated we get the donation prefilled in the "Send" tab, but obviously
we cannot do that transaction since the wallet is empty. However,
since the user chose to donate, maybe they are just about to send BTC
to this wallet? If so, once they are received they can go back to the
"Send" tab later and finish their donation, so it still makes some
sense.