Stop installing libudev-dev
It was newly installed in Tails 6.10 because the updated trezor package
Suggests: libudev-dev. That package is only needed to update the firmware of
older Trezor devices, but it's not sufficient: to do so one needs a bunch more
dev packages that we don't install, and technical skills. So that's not a good
enough reason to ship this package by default.