Improve test_partitioning_errors=part_resize
I.e. we simulate that only the repartitioning of the sgdisk command
failed. If we don't do this the GUIDs are not randomized, so
verify_partition_table will result in guid-not-randomized instead of
system-partition-not-resized due to how the checks are ordered. So an
alternative fix is to reorder those checks, but I think this more
accurate failure simulation is the better approach.