Re-import historical self-signatures on our signing key
GnuPG uses the current view of the public key (aka. certificate) to verify
signatures, but "When Sequoia checks a signature, it checks it using a view of
the certificate at the time the signature was allegedly created" (details:
https://gitlab.com/sequoia-pgp/sequoia/-/issues/758 and
https://openpgp.dev/book/adv/certificates.html#limitations-that-can-result-from-stripping-historical-self-signatures).

Whenever we update our OpenPGP signing key and export the updated version to
tails-signing.key in Git, it's possible that we strip historical
self-signatures, which can make previously generated UDF signatures invalid for
Sequoia. That's what happened in d97d4e9983f8c65c4e5cb1a66c56c771f265499c.

Without this change, for a bunch of our UDFs, this command would fail:

   sqop verify $UDF wiki/src/tails-signing.key < $UDF

… which would block updating our check-udfs-signature script.