Regularly merge stable into devel
Ideally we would merge master into stable, *then* stable into devel,
so that devel gets the latest commits from master every day,
just like stable does.

Unfortunately it's not trivial to achieve with the current implementation of
ensure-merge-request: the merge of master into stable is done in an asynchronous
manner, once the pipeline succeeds, so we can't order another GitLab CI job
after that happened solely using the GitLab CI configuration language.

So, with this simplistic commit, in the general case (no merge conflict,
pipeline succeeds), devel will lag behind master by 1 extra day compared to
stable. It's OK. It's much better than "until a developer thinks of doing this
merge manually". Surely we could teach ensure-merge-request to take care of the
whole merge flow, and create the second MR once the first one has been merged,
but at this stage it does not seem to be worth the effort. We can always come
back to this later if this limitation of the first iteration is
sufficiently problematic.

refs #18003