Test suite: Bump the time we wait before checking memory poisoning, again
Waiting 3 seconds instead of 1, as we did in !1630, helped a lot:
tails/tails#20354 used to happen "often", now it's quite rare. But it still
happens: I've seen "3.766% of the free memory still has the pattern, but less
than 0.800% was expected".

In
https://gitlab.tails.boum.org/tails/tails/-/merge_requests/1630#note_240453
@anonym wrote:

> 10 sounds a bit much from a "what latency we call good enough for this
> feature" perspective, not from a "minimizing test suite failures" one. I think
> your original 3 is worth trying.

So let's try to find a middle ground.

Note that as opposed to !1630, which bumped the sleep time from 1s to 3s for all
the memory erasure scenarios, here I'm bumping it only for the specific scenario
where I've seen a recent failure. Interestingly, while tails/tails#20354 does
not say which exact scenarios were found to be affected back then, in the
attached debug.log the only affected one is the scenario that I saw failing
here. So it might be than 1 second worked just fine for all the other scenarios,
and only this one needed a bump. I'm not going to dive into this deeper as
3 seconds seems to be a reasonable baseline.

For more context, see the full discussion around the aforementioned comment.