TCA: avoid race condition
The race is between Tor bootstrapping and the lock file being created,
so if bootstrapping finishes before the lock file is created, then the
lock file wouldn't do what it is supposed to in
tor_wait_until_bootstrapped. It is probably very unlikely or even
practically impossible for this to happen, but let's not take any
risks.