Remove xdg-desktop-portal-gnome
The Gnome FileChooser portal has many issues when interacted with
through Dogtail, which would require ugly hacks in the automated test
suite. By removing it we get the GTK FileChooser which does not have
these issues. So far the only place it will be used is in Tor Browser,
whose native FileChooser is identical to the GTK one any way, so less
surprises for users, which is nice.

In Trixie xdg-desktop-portal-gnome works completely differently and
integrates with Nautilus, and do not have these issues, so we can
switch to it then.

Also, it's worth noting that I first tried to keep
xdg-desktop-portal-gnome and instead configure which portal to use for
which interfave through ~/.config/xdg-desktop-portal/portals.conf like
this:

    [preferred]
    default=gnome
    org.freedesktop.impl.portal.FileChooser=gtk

but for some reason (too old version of xdg-desktop-portal?) that just
doesn't work, although it should according to portals.conf(5).