Tor Browser: re-enable namespace isolation for the sandbox
Tor Browser's sandbox uses unprivileged namespaces as part of its
sandbox, but flatpak breaks this through its seccomp filter which
restricts the needed syscall. For this reason we have patched flatpak
to allow disabling the seccomp filter, which is what this commit
does. Note that Tor Browser already runs its own, tailored seccomp
filter, so we're not worse off compared to running Tor Browser
"natively".

This is still not ideal, however, since it weakens defenses in the
security boundary between the flatpak and the host system in the event
Tor Browser is compromised, but we decided that we prefer full
operation of Tor Browser's sandbox over that. However, Firefox is in
the process of introducing a proper fix for this so we can drop this
patch: https://bugzilla.mozilla.org/show_bug.cgi?id=1756236