tails-documentation: don't start Tor Browser via .desktop file
The org.boum.tails.TorBrowser desktop file uses the "@@u ... @@"
magic, which would make it so the HTML document passed as argument is
shared using the doc portal. But then _only_ the HTML document is
shared, not any of the other HTML files it links to, images, css,
etc. so we need to force flatpak to not use the doc portal by avoiding
the "@@u ... @@" magic.

Details: https://gitlab.tails.boum.org/tails/tails/-/merge_requests/1025#note_254087