Consider the device corrupt if sgdisk detects corruption during first boot (refs: tails#20797)
We have seen one instance (wb:17cdc07f4a6f6c7c2a3cd5a8b6a51f6c) where
sgdisk reported CRC mismatches for the backup GPT header and backup
partition header even before we did any repartitioning. Once the
repartitioning was done it actually "fixed" the issue so it was not
detected in verify_partition_table, but the device was still not
working properly because later when the user tried to create a
Persistent Storage that failed during `cryptsetup luksFormat` with:

    udisksd: Error getting '/dev/sda2' metadata_size: Failed to load
    device: Invalid argument (g-bd-crypto-error-quark, 0)

So let's catch this issue earlier and consider the device corrupt.