Flatpak runtime: don't do anything special during early patching
This only made sense when the runtime root was built from the SquashFS
(i.e. prior to the "Mount / into flatpak runtime instead of the
SquashFS" commit), now that it's built from the usual root filesystem
the effects of early patching will end up in the runtime without
special handling.

(Note that there is a difference: for the Flatpak runtime we mounted
config/chroot_local-includes/usr on top of /usr as the lowerdir in the
overlay, so new/deleted files would becomes live, which is not the
case with early patching's `bind-new-includes`, which only affects
files changed at boot time. But there is no reason the runtime should
get special treatment, if we ever want early patching to use an
overlay mount approach instead of per-file bind-mounts we'll fix it
everywhere, i.e. for /.)