Tor Browser: fix the local page detection vs "@@u ... @@"
Notably, this also moves the logic for handling local pages into the
tor-browser script so it's solved everywhere, which goes nicely with
the plans to deprecate tails-documentation (tails#20144). But since
the logic is moved from tails-documentation it also loses the control
needed for its --force-local, which was only use by TCA, but it's no
more relevant there than anywhere else, so we drop it.