Refactor tails-documentation: avoid converting URLs twice
Previously, invoking tails-documentation with a URL as an argument would
have resulted in this link being converted to a relative path, then
converted to an URL, which would then be passed to tor-browser.
Now, this is short-circuited.

Invoking tails-documentation with a relative path as an argument would
still result in a double conversion: the relative path is converted to a
URL, which is passed to tor-browser, which, if we're online, might
convert it back to an absolute path.