Unsafe Browser: remove start notification
It made sense years ago when the setup was way less efficient and the
starting could confuse users. These days the difference between how
long it takes to start Tor Browser is about 1 second on my system, so
I don't see the need for this notification any longer.

(Note that what made me even think of this in the first place is that
when migrating to Tor Browser 14.5 (refs: tails#20904) this
notification started interfering in "Scenario Outline: Tails is
localized for every tier-1 language" by staying visible much
longer. Weirdly enough it didn't do this in other scenarios like those
in unsafe-browser.feature. So by dropping this notification, which
makes sense in itself, I could avoid spending time investigating this
uninteresting problem.)