Don't ask for Wi-Fi adapters anymore
We've had this note on the website since 2018 and it has never helped identify
a reliable Wi-Fi adapter that people can buy easily in brick-and-mortar shops
around the world.

In my experience, common hardware manufacturers (TP-Link and D-Link, for
example) release new models too often. The new ones don't work with Linux. In
the unlikely even where one of them starts working with Linux, it's not
produced by the manufacturer anymore. Also, different models are available in
different countries.

In comparison, the 2 adapters that we recommend for online shopping have been
reliable since then.

Too bad for people who live in places where Amazon doesn't operate, but let's
simplify our installation instructions and stop wasting everybody's time on
this.