Enable automated-testing@tails.net for the GDM session later
Enabling GNOME Shell extensions requires the
org.gnome.Shell.Extensions D-Bus services to be running, and they it
takes a while (due to its own dependencies probably). In the previous
location this command blocked the Welcome Screen from starting for 36
seconds (on my system), but by moving it to the new place this is
changed to a wait of just 18 seconds without blocking the Welcome
Screen at all, so the total reduced delay is possibly even higher.