Handle reader connection reset in tca-portal
The tca-portal handler might endlessly wait for a reader message even
if the reader is not running anymore.

Such a situation can occur if the connection the reader is reading from
is reset by its client. The handler will never leave its loop in such a
case.

The amount of parallel running handlers is limited. Keeping these
handlers from stopping could lead to a local denial of service of the
tca-portal service.