Reduce amount of zombies in tca-portal
The handle_request method in Handler class starts a process without
ever joining with it again. Ignore the SIGCHLD signal to avoid the
creation of zombies in the first place.

Reset the SIGCHLD handling before spawning external processes to not
interfere their child handling.

This still leads to a zombie (runas) as long as a sub process like
tor browser is running. This is okay though because the file
descriptors are passed along and might still be used for eventual
status return value.