Revert "acng can cache broadcom firmwares"
This reverts commit 0cf17b3bba58e27a91a43982acbbf6dd51d8bd0f.

To make it work, we would need to have firmware-b43-installer download
files fom http://www.lwfinger.com (http, not https). And to do that, we
would need to patch the package.
Now, we *are* indeed patching the package, so we could very easily make
it work. But I hope we can get rid of it, and go back to using Debian
packages, which would break this improvement again.