Move the Tails Flatpak runtime and fake Flatpaks to its own installation
We already have users that install flatpak themselves and make
/var/lib/flatpak/app persistent, which means that any future Tails
releases' changes to our Flatpaks' metadata will be lost for those
users (refs: tails#21016). We avoid such issues by moving our runtime
and fake Flatpaks to a custom installation. This separation will
likely solve similar and other issues once we want to add proper
support for installing Flatpaks in Tails.

Unfortunately, due to a bug in Flatpak, overrides do not respect the
--installation option used to implement this:

    https://github.com/flatpak/flatpak/issues/6256

This means that if a user makes all of /var/lib/flatpak persistent,
then any future Tails releases' changes to
/var/lib/flatpak/overrides/org.boum.tails.TorBrowser will be lost in
the same way. Once that bug is fixed we will move those overrides to
the custom installation:

    git mv config/chroot_local-includes/var/lib/flatpak/ \
           config/chroot_local-includes/var/lib/flatpak-tails/