Test suite: search for "Start Tails" button with support for localization
In the `greeter()` method we pass `translation_domain: 'tails'` when
initializing the Welcome Screen Dogtail::Application, so all "names"
will be localized with `translate()` automatically when we search with
`.child()` et al. However, the PO translations include accelerators,
so they need to be included in order to find the correct
translation. (And the translated string likely includes an
accelerator, e.g. "Tails _starten" for German, which Dogtail wouldn't
find, so we have to drop it, but `translate()` does that by defaulting
to `drop_accelerator: true`.)

This will fix the `Persistent Greeter options` scenario which got
broken due to commit b7fda4855409b2c726b1329196f593af98159ee9.