Anticipate the fix for flatpak#6256
The upstream flatpak#6256 issue is "--installation is not respected
for overrides". In Tails that means that the org.boum.tails.TorBrowser
overrides file cannot live in the /var/lib/flatpak-tails installation
where its Flatpak lives, but must be put in the default
/var/lib/flatpak installation that users also might use. So we use a
symlink so it's present in both locations, so Tails will seamlessly
switch to the fixed location once we install a flatpak with the fix.

Refs: tails#21016, tails#21060