Simplify
Don't imply that `bridge` bridges can be used. They are barely used these days.
The UI still allows using them, but I don't think it's worth making the
documentation more complicated to read because of them.

The notion of "discreet" actually depends on the context: some censors might be
able to detect some bridge types and not others.