make ruff faster by spawning much less processes
is-file-type provides a filter subcommand which is very good at doing
exactly what we want it to do, in a way which is very friendly to pipes.

When you have lots of changed files (ie: your branch diverged a lot from
origin/stable) this can bring 5.4s to 0.1s (anecdotal example).