Install libdrm-nouveau2 and libdrm-radeon1
In bookwork we automatically installed these since they were
dependencies for libgl1-mesa-dri, but in trixie that is no longer the
case so we have to explicitly install them. While limited testing
shows that direct rendering works without these for at least the
nouveau driver they're small enough to include any way to stay on the
safe side.