Don't include hunspell dictionaries in the bdic format
These dictionaries are shipped *in addition to* the same dictionaries
in the .dic + .aff format and grow the image size significantly.

I understand only Chrome & friends, as well as apps that use Qt WebEngine, use
the bdic format, so no app we ship should lose spellchecking if we exclude
those.

refs tails/tails#21006