Install rtkit
On #19627 (comment 265688) we noticed there's a lot of noise about
"org.freedesktop.RealtimeKit1" does not exist in the Journal. pipewire-bin
Recommends rtkit: it's tiny, it'll make the Journal quieter, and it's probably
a good idea to follow the Debian & upstream defaults here, to ensure audio does
not crackle when the system is under load.

Note: if we only wanted to make the Journal quieter, chances are that setting
DISABLE_RTKIT via an override for the relevant pipewire unit is enough.

refs tails/tails#21119