Skip the onboarding of Inkscape
When opening Inkscape in Tails, people have to click through 3 screens of
onboarding. They are fine outside of Tails, where people only have to go through
them once. But, in the context of Tails, I think that they do more harm than
good. So let's skip the first 2 screens. The last screen is more useful as it
allows to choose a drawing template.

refs tails/tails#21091