Test suite: remove SFTP automated test
In d9e14385 the "I connect to an SFTP server on the Internet" test scenario
switched from using Nautilus to using the Places menu, "because the former is
difficult to interact with using Dogtail since Trixie". And this MR removes the
Places menu.

So I'm dropping the test because I think it has a pretty bad cost/benefit ratio.

For details, see https://gitlab.tails.boum.org/tails/tails/-/merge_requests/2362#note_266113