needed-package-updates: don't track proposed-updates
The Debian $dist-proposed-updates APT suite is used to host *candidate*
packages, which may or may not be accepted for the upcoming Debian stable
point-release. We never enable it on our release branches because it's not
considered stable enough for production use.

So the course of action that our CI is currently suggesting is not appropriate:

  INFO: Need to upgrade to a new APT snapshot:
    - b43-fwcutter (1:019-8tails1) to Debian bookworm-proposed-updates (1:019-8+deb12u1)