needed-package-updates: only warn when we're newer than Debian
Currently for the tor package: Debian only has 0.4.8.16-1, while we install
0.4.8.17-1~d12.bookworm+1 from deb.torproject.org, so the script is raising
an error.

In general, I think it's acceptable that we ship a package that's newer than
Debian: this can happen in other legit cases, e.g. if we need to fork the
version from sid. In that case, we'll need another way to track sid to ensure
our fork is up to date. I don't think needed-package-updates trying to forbid
this situation from happening at all is particularly helpful.

Still, a warning can be a good safety net: next time needed-package-updates
fails for another reason, someone may spot a problem if there is actually one.
If this causes alert fatigue, we should drop this warning.

And while I'm it, I'm fixing the error message: "suite" is the last checked
suite, which is not necessarily the one that has the package at
"suite_version". E.g. in the case at hand, the error message was:

  ERROR: tor: the package version (0.4.8.17-1~d12.bookworm+1) is higher than the version on experimental (0.4.8.16-1)

… which is incorrect: 0.4.8.16-1 is in sid, not in experimental.