Release process: avoid potential mismatch for build reproducibility check
Since f605190e423e792a3382006180936ce2ac0b21d1 we are telling the RM to set
matching_jenkins_images_build_id to the build we *think* should reproduce our
local build. But we were not actually comparing this build with our local build.
Instead, we were comparing the last successful build, which may or may not be
the same.

Let's fix that by ensuring we compare the same build that we'll then encode as
the matching one.