report all errors
If we don't, it's very easy for the RM (or whoever calls such a script)
to think that the error is the *only* error we have. In a recent case,
the only displayed error was about a very, very old Tails release. It
wouldn't be surprising if a RM thinks "oh, the only error is about a 10
years old release, who cares, let's proceed".