Bump memory for Trixie builds on Jenkins
This number used to work fine, but then we based this MR on trixie, and
it FTBFS.

Now, there is not much point in getting a "sharp" number here: our
isoworkers have 20GBs, and this number is only useful for Jenkins
builds.

So let's just give it a large bump.