Update inlines: drop currentlang() for pages that are not translatable anymore
For backwards compatibility, the index pages and RSS/Atom feeds remain
localized, but instead of listing only contents in the same language,
they're now all listing contents in English.

It would probably be possible to stop localizing the index pages too, but it's
tricky to do that with a broad variety of clients that may or may not honor
HTTP redirects.