Set system locale with update-locale
In commit a99644c348378057ec371dcc422909486173f961 we removed this,
but at the time it was done with `localectl set-locale` and apparently
did nothing but spam errors to the journal. But now
/etc/locale.conf (and the deprecated /etc/default/locale) is not
updated according to what is set in Welcome Screen, which is
problematic for our scripts that source that file to learn the
locale. So we reintroduce what was removed in that commit, but using
update-locale instead, which does not pollute the journal.

Refs: tails/tails#21118