refactor building https-get-expired to DRY
similar code was present in two different areas. And in fact, there were
slight, unintended differences: ldflags were not customized and the
binary was not stripped. While this was not bad in itself (those changes
are there for reproducibility), the more changes we introduce between CI
and product, the more difficult it is to attest quality and debug
problems.

Now this is in a central place.