skip the onboarding of GIMP
refs #21142

The reason we're not putting this in /etc/gimp/ is because files in
/etc/ are merged with files in $HOME/.config/.
This might have unintened consequences for people persisting
~/.config/GIMP/. In fact, when we'll ship an updated version of GIMP,
we'll need to bump config-version. But this would have the effect of
also marking the user-controlled file in ~/.config/GIMP/ as ready for
the new version, which it might not be.

See https://gitlab.tails.boum.org/tails/tails/-/merge_requests/2450#note_268909