Restart systemd-localed.service to ensure org.freedesktop.locale1 updates
... and emits the signal which tpsd then picks up and uses to change
which language it translates its strings to (refs tails#21210).

update-locale only writes to /etc/locale.conf, so it is not enough to
make org.freedesktop.locale1 aware of the new locale, for that you
also need to restart systemd-localed.service. This is seemingly the
only way to do this because on Debian due to the D-Bus policy
/usr/share/dbus-1/system.d/systemd-localed-read-only.conf which
explains itself with this comment:

    On Debian and derivatives keymap/locales/etc are not set via localed,
    but from legacy and incompatible components. But we still need to
    enable localed so that GNOME can query it. Ensure not even root can
    use it to modify the settings.

and then goes ahead and prevents all users from calling any of the
setter methods (used by e.g. `localectl set-locale`).