Hide the GnuPG Log Viewer from the Apps menu
It's new in Tails 7.0~rc2, and it's also broken because it relies on watchgnupg,
which comes from gnupg-utils package, which we installed on 6.x, but don't
install anymore on Trixie.

It would be more involved to hide the same launcher from the Kleopatra
Tools menu, so I'm not even trying.

refs tails/tails#21214