Test suite: adjust to current systemd message
The new message is:

  System time advanced to built-in epoch: Tue 2025-09-09 08:38:33 UTC

I find no trace of the old message on codesearch.debian.net, so I'm going to
assume we won't get it anymore, and can fully replace it with the new one.

Also, this message can't be found with the _PID=1 match: it's emitted during
early boot, before journald starts, so it's collected by the kernel log buffer
and collected by journald once it starts later, so the matches available to us
here are the syslog ones, not the systemd ones.