don't restart htpdate if it already succeeded
refs #21014

If we already synced time, there is no reason for us to do it again;
but doing it again can *break* stuff, because our test suite will expect
this run to be successful, and will raise HtpdateError otherwise

There is no reason for doing that.
This doesn't solve the root cause, but should fix the part of the bug
which is visible in the test suite.