Release process: have CI guard the merge of the website release branch
During the 7.0 release process, a typo in the release notes, introduced in
web/release-7.0, that had not been caught beforehand, made its way into the
master branch once the RM manually merged web/release-7.0 into master.
Additionally, codespell detected another issue in the release notes,
which was actually a false positive.

This made CI fail, which blocked the deployment of the release to the live
website:

  https://gitlab.tails.boum.org/tails/tails/-/pipelines/19794

Then the RM had to notice this, fix the typo, figure out how to add a codespell
exception for the false positive, and push again. All this delayed the release.

Let's instead:

1. Do the web/release-$VERSION work in a merge request, so the release manager,
   technical writer, and innocent bystanders have greater chances to notice any CI
   failure earlier, outside of the critical path to the release.

2. Have CI guard the merge of the website release branch, so that regressions
   don't land undetected on the master branch.