Release notes: replace unrealistic expectation with hopefully more realistic one
The requirement to create a MR was added in 2021. I can't recall it being
followed recently. Too bad, it would have been useful at least once. But it also
adds quite some overhead, that we don't need anymore, thanks to the changes
brought by the previous commit.

So, let's as technical writers, instead of creating a MR of their own, to check
CI results for their work on the existing MR.