Release process: fix .packages comparison for major releases
We only want to compare against the last *stable* release for a first
release candidate, for subsequent release candidates we want to
compare against the previous one, and for the final release against
the last release candidate. By definition we get that by using the
PREVIOUS_RELEASE variable.

This still isn't ideal for a first release candidate of a (N+1).0
release, since comparing some N.x vs (N+1).0 will list basically every
single package due to the Debian major version bump. But we are not
regressing in this case.