Root Console policy logic is moved to Python
As suggested by anonym, we don't really need to ship two versions of the
same file: it's enough to ship one version, and then change it via
polkit-policy-change-message.

We could have left the _message in the policy file, but with this change
we are now centralizing the logic; it will also make it easy for
developers to run tests, because running polkit-policy-change-message is
enough, which wouldn't be the case otherwise.