Adapt to "Account Hub"
It's not the same assistant when you create a second account, but a new
"Account Hub" assistant.

The questions are very similar but the UI it totally different. I
suspect that both will converge at some point, so let's drop this note
from our doc for now.

The new button "New Account" (in blue) makes it easier to create a
second account.