use context managers where possible
that's more idiomatic, but won't change any real behavior.

The only "real" bug it solves is that, without this, the file `f` is not
closed. But even that's only true in Python implementations which are not
CPython, because in CPython, when a file object gets garbage-collected,
it's also closed. We have no plans to use a different Python
implementation, so that's just theoretical.