test suite web server sets content-type for HTML
We have seen a case in which the automated test suite was displaying the
page as plain text.
The previous commit ("fix HTML syntax error"), should be enough for the
browser to detect the type automatically, but let's also set the
Content-Type.
I have manually tested that this commit is enough to fix the problem
even without "fix HTML syntax error".