Remove dead code
We've removed the only user of this code in 7.1.

In the other place where we need to tell whether our website's ikiwiki supports
a given language (in the resolve method of tailslib.website Python library), we
did not use it anymore already: instead we check if the localized page exists in
the bundled website copy, which works fine.

If we ever need this facility again in shell code, it should be easy enough to
check it like that too — or to revert this commit.