Test suite: increase size of the 'live_hd' temporary medium to 2500 MiB
With the release of 7.x, the Tails ISO image grew to 1.87 GiB. The
scenario `Tails ignores a swap volume and another Tails that are on
an internal hard drive` copies the Tails ISO to a 2 GiB virtual device
which is only slightly larger than the ISO. A small increase in the
size of the ISO will cause an out-of-space write error and make the
scenario fail. To future-proof the scenario, increase the size of the
virtual device to 2500 MiB (2.44 GiB).