create-usb-image-from-iso: fix confusion between image size and system partition size
Until now, the code and comments pretended that the size of the system partition
was size of the ISO + 10 MiB. This was not the case. Instead, we were:

1. creating an image of size of the ISO + 10 MiB
2. creating a GPT inside this image
3. creating the GPT system partition inside this image,
   asking udisks to make it size of the ISO + 10 MiB,
   which is obviously impossible as the GPT itself needs space.

Let's instead rename all this to make it clear that "size of the ISO + 10 MiB"
is the size of the USB image, not the one of the system partition. And ask
udisks to make the system partition as large as it can, instead of asking it to
do impossible things.