Ensure fatresize does not block the boot with an interactive question
Without this, since we upgraded our Vagrant box to Trixie, during boot fatresize
gets stuck asking the user a question (that's hidden behind the splash screen,
and anyway can't be answered anymore):

  Warning: Not all of the space available to /dev/sda appears to be used, you
  can fix the GPT to use all of the space (an extra 2014 blocks) or continue
  with the current setting?

  Fix/Ignore:

That's because the "last usable block" encoded by sgdisk when it repartitions is
closer to the beginning of the disk than where libparted thinks it could
actually be.

For more details, see the discussion that starts on
https://gitlab.tails.boum.org/tails/tails/-/merge_requests/2543#note_272099

Another option would be to run this between the time we repartition with sgdisk,
and the time we run fatresize:

  parted --script --fix "${PARENT_DEVICE}" print free

This should make libparted happy enough with the disk layout, which should allow
fatresize to do its job without asking questions. But it feels a bit more
invasive than simply asking fatresize to not ask questions, so let's first
try this.