Test suite: consistently use a 7200 MiB disk when installing our USB image
Everywhere else that uses "I write the Tails USB image to disk", we use a 7200
MiB disk. Let's do the same here, to ensure that the different behavior we want
to verify here is caused by the disk being SATA, and not by the disk
being smaller.

We run this feature first (see `prioritized_features`) so this should not
increase the peak disk usage of a full test suite run.