Test suite: run xfwm4 in the test suite's X virtual framebuffer
So far we haven't been running any window manager at all, which has
worked fine until Debian Bookworm. But with virt-viewer >= 10.0 from
Debian Trixie this results in it not hiding its menu bar in fullscreen
mode which makes many of our images not match due to scaling
differences, and hides part of the screen.

Refs: tails#19064