Test suite: assert that a window manager is running before starting virt-viewer
The way we start the window manager (xfwm4) doesn't tell us whether it
fails or not, and there has already been issues with that, so let's be
sure it at least is running when starting virt-viewer. It would also
be a problem if it quits/crashes while virt-viewer is running, but
monitoring for that would require a lot more effort.

Refs: tails#19064