change test image for the Tails Greeter
the previous image was looking for text. but that text can be localized,
and given the work we're doing on tails#5501, there is no guarantee
anymore that that string will be in English. So let's just look for some
other part of the window which is not affected by l10n.

Let's keep in mind that "Disk partitioning errors without a persistent partition"
is "challenging" wrt to this change: relevant settings icons are greyed
out on purpose there.