Use 24-hour time format
This is what we used in 6.x, which resulted in 24-hour format both in the
desktop and on the lock screen. But in 7.x this setting switched to AM/PM, so
the time is displayed in 24-hour format on the desktop, but in 12-hour format in
the lock screen, which is confusing.

refs tails/tails#21342